PYIC-3030 Moved VerifiableCredentialJwtValidator to lib as needed by CiStorageService.
